modified bitmap & tsquare

+++ BitMap.cpp
@@ -19,11 +19,11 @@
 #include "vajolet.h"
 #include "BitMap.h"
 
-BitMap BitMap::RANKMASK[ tSquare::squareNumber ];			//!< bitmask of a rank given a square on the rank
-BitMap BitMap::FILEMASK[ tSquare::squareNumber ];			//!< bitmask of a file given a square on the rank
+BitMap BitMap::RANKMASK[ static_cast<int>(tSquare::squareNumber) ];			//!< bitmask of a rank given a square on the rank
+BitMap BitMap::FILEMASK[ static_cast<int>(tSquare::squareNumber) ];			//!< bitmask of a file given a square on the rank
 BitMap BitMap::SQUARECOLOR[ 2 ] = { 0ull, 0ull };
-BitMap BitMap::SQUARES_BETWEEN[ tSquare::squareNumber ][ tSquare::squareNumber ];
-BitMap BitMap::LINES[ tSquare::squareNumber ][ tSquare::squareNumber ];
+BitMap BitMap::SQUARES_BETWEEN[ static_cast<int>(tSquare::squareNumber) ][ static_cast<int>(tSquare::squareNumber) ];
+BitMap BitMap::LINES[ static_cast<int>(tSquare::squareNumber) ][ static_cast<int>(tSquare::squareNumber) ];
 
 /*	\brief return a string representing a bitmap
 	\author Marco Belli
@@ -65,8 +65,8 @@ std::string BitMap::to_string() const
 */
 void BitMap::init(void)
 {
-	BitMap DIAGA1H8MASK[ tSquare::squareNumber ];
-	BitMap DIAGA8H1MASK[ tSquare::squareNumber ];
+	BitMap DIAGA1H8MASK[ static_cast<int>(tSquare::squareNumber) ];
+	BitMap DIAGA8H1MASK[ static_cast<int>(tSquare::squareNumber) ];
 	
 	SQUARECOLOR[0] = 0ull;
 	SQUARECOLOR[1] = 0ull;
@@ -79,51 +79,51 @@ void BitMap::init(void)
 		//initialize 8-bit rank mask
 		//===========================================================================
 
-		RANKMASK[ sq ] = getSquareFromFileRank( tFile::A, rank ) ;
-		RANKMASK[ sq ] += getSquareFromFileRank( tFile::B, rank );
-		RANKMASK[ sq ] += getSquareFromFileRank( tFile::C, rank );
-		RANKMASK[ sq ] += getSquareFromFileRank( tFile::D, rank );
-		RANKMASK[ sq ] += getSquareFromFileRank( tFile::E, rank );
-		RANKMASK[ sq ] += getSquareFromFileRank( tFile::F, rank );
-		RANKMASK[ sq ] += getSquareFromFileRank( tFile::G, rank );
-		RANKMASK[ sq ] += getSquareFromFileRank( tFile::H, rank );
+		RANKMASK[ static_cast<int>(sq) ] = getSquareFromFileRank( tFile::A, rank ) ;
+		RANKMASK[ static_cast<int>(sq) ] += getSquareFromFileRank( tFile::B, rank );
+		RANKMASK[ static_cast<int>(sq) ] += getSquareFromFileRank( tFile::C, rank );
+		RANKMASK[ static_cast<int>(sq) ] += getSquareFromFileRank( tFile::D, rank );
+		RANKMASK[ static_cast<int>(sq) ] += getSquareFromFileRank( tFile::E, rank );
+		RANKMASK[ static_cast<int>(sq) ] += getSquareFromFileRank( tFile::F, rank );
+		RANKMASK[ static_cast<int>(sq) ] += getSquareFromFileRank( tFile::G, rank );
+		RANKMASK[ static_cast<int>(sq) ] += getSquareFromFileRank( tFile::H, rank );
 
 		//===========================================================================
 		//initialize 8-bit file mask
 		//===========================================================================
-		FILEMASK[ sq ] = getSquareFromFileRank( file, tRank::one);
-		FILEMASK[ sq ] += getSquareFromFileRank( file, tRank::two);
-		FILEMASK[ sq ] += getSquareFromFileRank( file, tRank::three);
-		FILEMASK[ sq ] += getSquareFromFileRank( file, tRank::four);
-		FILEMASK[ sq ] += getSquareFromFileRank( file, tRank::five);
-		FILEMASK[ sq ] += getSquareFromFileRank( file, tRank::six);
-		FILEMASK[ sq ] += getSquareFromFileRank( file, tRank::seven);
-		FILEMASK[ sq ] += getSquareFromFileRank( file, tRank::eight);
+		FILEMASK[ static_cast<int>(sq) ] = getSquareFromFileRank( file, tRank::one);
+		FILEMASK[ static_cast<int>(sq) ] += getSquareFromFileRank( file, tRank::two);
+		FILEMASK[ static_cast<int>(sq) ] += getSquareFromFileRank( file, tRank::three);
+		FILEMASK[ static_cast<int>(sq) ] += getSquareFromFileRank( file, tRank::four);
+		FILEMASK[ static_cast<int>(sq) ] += getSquareFromFileRank( file, tRank::five);
+		FILEMASK[ static_cast<int>(sq) ] += getSquareFromFileRank( file, tRank::six);
+		FILEMASK[ static_cast<int>(sq) ] += getSquareFromFileRank( file, tRank::seven);
+		FILEMASK[ static_cast<int>(sq) ] += getSquareFromFileRank( file, tRank::eight);
 		
 		//===========================================================================
 		//initialize SQUARECOLOR
 		//===========================================================================
-		SQUARECOLOR[ getColor(sq) ] += sq;
+		SQUARECOLOR[ static_cast<unsigned int>( getColor(sq) ) ] += sq;
 		
 		//===========================================================================
 		//Initialize 8-bit diagonal mask
 		//===========================================================================
 		
-		DIAGA1H8MASK[sq] = 0ull;
-		DIAGA8H1MASK[sq] = 0ull;
-		int diaga8h1 = file + rank; // from 0 to 14, longest diagonal = 7
+		DIAGA1H8MASK[ static_cast<int>(sq) ] = 0ull;
+		DIAGA8H1MASK[ static_cast<int>(sq) ] = 0ull;
+		int diaga8h1 = static_cast<int>(file) + static_cast<int>(rank); // from 0 to 14, longest diagonal = 7
 		if (diaga8h1 < 8)  // lower half, diagonals 0 to 7
 		{
 			for (int square = 0 ; square <= diaga8h1 ; square ++)
 			{
-				DIAGA8H1MASK[ sq ] += getSquareFromFileRank( tFile( square ), tRank( diaga8h1 - square ) ) ;
+				DIAGA8H1MASK[ static_cast<int>(sq) ] += getSquareFromFileRank( tFile( square ), tRank( diaga8h1 - square ) ) ;
 			}
 		}
 		else  // upper half, diagonals 8 to 14
 		{
 			for (int square = 0 ; square < 15 - diaga8h1 ; square ++)
 			{
-				DIAGA8H1MASK[ sq ] += getSquareFromFileRank( tFile( diaga8h1 + square - 7 ), tRank( 7 - square ) );
+				DIAGA8H1MASK[ static_cast<int>(sq) ] += getSquareFromFileRank( tFile( diaga8h1 + square - 7 ), tRank( 7 - square ) );
 			}
 		}
 
@@ -131,19 +131,19 @@ void BitMap::init(void)
 		//===========================================================================
 		//Initialize 8-bit diagonal mask, used in the movegenerator (see movegen.ccp)
 		//===========================================================================
-		int diaga1h8 = file - rank; // from -7 to +7, longest diagonal = 0
+		int diaga1h8 = static_cast<int>(file) - static_cast<int>(rank); // from -7 to +7, longest diagonal = 0
 		if (diaga1h8 > -1)  // lower half, diagonals 0 to 7
 		{
 			for (int square = 0 ; square <= 7 - diaga1h8 ; square ++)
 			{
-				DIAGA1H8MASK[ sq ] += getSquareFromFileRank( tFile( diaga1h8 + square ), tRank( square ) );
+				DIAGA1H8MASK[ static_cast<int>(sq) ] += getSquareFromFileRank( tFile( diaga1h8 + square ), tRank( square ) );
 			}
 		}
 		else
 		{
 			for (int square = 0 ; square <= 7 + diaga1h8 ; square ++)
 			{
-				DIAGA1H8MASK[ sq ] += getSquareFromFileRank( tFile(square), tRank( square - diaga1h8 ) );
+				DIAGA1H8MASK[ static_cast<int>(sq) ] += getSquareFromFileRank( tFile(square), tRank( square - diaga1h8 ) );
 			}
 		}
 	}
@@ -152,8 +152,8 @@ void BitMap::init(void)
 	{
 		for( auto sq2: tSquareRange() )
 		{
-			LINES[ sq1 ][ sq2 ] = 0ull;
-			SQUARES_BETWEEN[ sq1 ][ sq2 ] = 0ull;
+			LINES[ static_cast<int>(sq1) ][ static_cast<int>(sq2) ] = 0ull;
+			SQUARES_BETWEEN[ static_cast<int>(sq1) ][ static_cast<int>(sq2) ] = 0ull;
 			
 			if( sq1 != sq2 )
 			{
@@ -161,13 +161,13 @@ void BitMap::init(void)
 				{
 					// stessa colonna
 
-					LINES[ sq1 ][ sq2 ] = FILEMASK[ sq1 ];
+					LINES[ static_cast<int>(sq1) ][ static_cast<int>(sq2) ] = FILEMASK[ static_cast<int>(sq1) ];
 					if( getRank( sq2 ) > getRank( sq1 ) ) // in salita
 					{
 						tRank temp = getRank( sq1 ) + 1;
 						while( temp < getRank( sq2 ) )
 						{
-							SQUARES_BETWEEN[ sq1 ][ sq2 ] += getSquareFromFileRank( getFile( sq1 ), temp );
+							SQUARES_BETWEEN[ static_cast<int>(sq1) ][ static_cast<int>(sq2) ] += getSquareFromFileRank( getFile( sq1 ), temp );
 							temp++;
 						}
 					}
@@ -176,20 +176,20 @@ void BitMap::init(void)
 						tRank temp = getRank( sq1 ) - 1;
 						while( temp > getRank( sq2 ) )
 						{
-							SQUARES_BETWEEN[ sq1 ][ sq2 ] += getSquareFromFileRank( getFile( sq1 ), temp );
+							SQUARES_BETWEEN[ static_cast<int>(sq1) ][ static_cast<int>(sq2) ] += getSquareFromFileRank( getFile( sq1 ), temp );
 							temp--;
 						}
 					}
 				}
 				if( getRank( sq1 ) == getRank( sq2 ) )
 				{
-					LINES[sq1][sq2] = RANKMASK[sq1];
+					LINES[ static_cast<int>(sq1) ][ static_cast<int>(sq2) ] = RANKMASK[ static_cast<int>(sq1) ];
 					if( getFile( sq2 ) > getFile( sq1 ) ) // in salita
 					{
 						tFile temp = getFile( sq1 ) + 1;
 						while( temp < getFile( sq2 ) )
 						{
-							SQUARES_BETWEEN[sq1][sq2] += getSquareFromFileRank( temp, getRank( sq1 ) );
+							SQUARES_BETWEEN[ static_cast<int>(sq1) ][ static_cast<int>(sq2) ] += getSquareFromFileRank( temp, getRank( sq1 ) );
 							temp++;
 						}
 					}
@@ -198,21 +198,21 @@ void BitMap::init(void)
 						tFile temp = getFile( sq1 ) - 1;
 						while( temp > getFile( sq2 ) )
 						{
-							SQUARES_BETWEEN[sq1][sq2] += getSquareFromFileRank( temp, getRank( sq1 ) );
+							SQUARES_BETWEEN[ static_cast<int>(sq1) ][ static_cast<int>(sq2) ] += getSquareFromFileRank( temp, getRank( sq1 ) );
 							temp--;
 						}
 					}
 				}
-				if( DIAGA1H8MASK[sq1].isSquareSet( sq2 ) )
+				if( DIAGA1H8MASK[ static_cast<int>(sq1) ].isSquareSet( sq2 ) )
 				{
-					LINES[sq1][sq2] = DIAGA1H8MASK[sq1];
+					LINES[ static_cast<int>(sq1) ][ static_cast<int>(sq2) ] = DIAGA1H8MASK[ static_cast<int>(sq1) ];
 					if( getFile( sq2 ) > getFile( sq1 ) ) // in salita
 					{
 						tFile temp = getFile( sq1 ) + 1;
 						tRank temp2 = getRank( sq1 ) + 1;
 						while( temp < getFile( sq2 ) )
 						{
-							SQUARES_BETWEEN[sq1][sq2] += getSquareFromFileRank( temp, temp2 );
+							SQUARES_BETWEEN[ static_cast<int>(sq1) ][ static_cast<int>(sq2) ] += getSquareFromFileRank( temp, temp2 );
 							temp++;
 							temp2++;
 						}
@@ -223,22 +223,22 @@ void BitMap::init(void)
 						tRank temp2 = getRank( sq1 ) - 1;
 						while( temp > getFile( sq2 ) )
 						{
-							SQUARES_BETWEEN[sq1][sq2] += getSquareFromFileRank( temp, temp2 );
+							SQUARES_BETWEEN[ static_cast<int>(sq1) ][ static_cast<int>(sq2) ] += getSquareFromFileRank( temp, temp2 );
 							temp--;
 							temp2--;
 						}
 					}
 				}
-				if(DIAGA8H1MASK[sq1].isSquareSet( sq2 ) )
+				if(DIAGA8H1MASK[ static_cast<int>(sq1) ].isSquareSet( sq2 ) )
 				{
-					LINES[sq1][sq2] = DIAGA8H1MASK[sq1];
+					LINES[ static_cast<int>(sq1) ][ static_cast<int>(sq2) ] = DIAGA8H1MASK[ static_cast<int>(sq1) ];
 					if( getFile( sq2 ) > getFile( sq1 ) ) // in salita
 					{
 						tFile temp = getFile( sq1 ) + 1;
 						tRank temp2 = getRank( sq1 ) - 1;
 						while( temp < getFile( sq2 ) )
 						{
-							SQUARES_BETWEEN[sq1][sq2] += getSquareFromFileRank( temp, temp2 );
+							SQUARES_BETWEEN[ static_cast<int>(sq1) ][ static_cast<int>(sq2) ] += getSquareFromFileRank( temp, temp2 );
 							temp++;
 							temp2--;
 						}
@@ -249,7 +249,7 @@ void BitMap::init(void)
 						tRank temp2 = getRank( sq1 ) + 1;
 						while( temp > getFile( sq2 ) )
 						{
-							SQUARES_BETWEEN[sq1][sq2] += getSquareFromFileRank( temp, temp2 );
+							SQUARES_BETWEEN[ static_cast<int>(sq1) ][ static_cast<int>(sq2) ] += getSquareFromFileRank( temp, temp2 );
 							temp--;
 							temp2++;
 						}
